backend/alpha_archives/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
ImageHash==4.2.1
mysql-connector==2.2.9
mysqlclient==2.0.3
numpy==1.21.2
Pillow==8.3.2
pytz==2021.1
PyWavelets==1.1.1
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
